+++ metrics-graphite/src/main/java/com/yammer/metrics/reporting/GraphiteReporter.java
@@ -1,10 +1,12 @@
 package com.yammer.metrics.reporting;
 
 import com.yammer.metrics.Metrics;
-import com.yammer.metrics.core.MetricsRegistry;
 import com.yammer.metrics.core.*;
+import com.yammer.metrics.core.VirtualMachineMetrics.*;
 import com.yammer.metrics.util.MetricPredicate;
 import com.yammer.metrics.util.Utils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.OutputStreamWriter;
@@ -14,22 +16,25 @@ import java.net.Socket;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import static com.yammer.metrics.core.VirtualMachineMetrics.daemonThreadCount;
+import static com.yammer.metrics.core.VirtualMachineMetrics.fileDescriptorUsage;
+import static com.yammer.metrics.core.VirtualMachineMetrics.garbageCollectors;
+import static com.yammer.metrics.core.VirtualMachineMetrics.heapUsage;
+import static com.yammer.metrics.core.VirtualMachineMetrics.memoryPoolUsage;
+import static com.yammer.metrics.core.VirtualMachineMetrics.nonHeapUsage;
+import static com.yammer.metrics.core.VirtualMachineMetrics.threadCount;
+import static com.yammer.metrics.core.VirtualMachineMetrics.threadStatePercentages;
+import static com.yammer.metrics.core.VirtualMachineMetrics.uptime;
 
-import static com.yammer.metrics.core.VirtualMachineMetrics.*;
 
 /**
  * A simple reporter which sends out application metrics to a
  * <a href="http://graphite.wikidot.com/faq">Graphite</a> server periodically.
  */
-public class GraphiteReporter implements Runnable {
-    private static final Logger log = LoggerFactory.getLogger(GraphiteReporter.class);
-    private final ScheduledExecutorService tickThread;
-    private final MetricsRegistry metricsRegistry;
+public class GraphiteReporter extends AbstractPollingReporter {
+    private static final Logger LOG = LoggerFactory.getLogger(GraphiteReporter.class);
     private final String host;
     private final int port;
     private final String prefix;
@@ -110,7 +115,7 @@ public class GraphiteReporter implements Runnable {
             final GraphiteReporter reporter = new GraphiteReporter(metricsRegistry, host, port, prefix, predicate);
             reporter.start(period, unit);
         } catch (Exception e) {
-            log.error("Error creating/starting Graphite reporter:", e);
+            LOG.error("Error creating/starting Graphite reporter:", e);
         }
     }
 
@@ -150,8 +155,7 @@ public class GraphiteReporter implements Runnable {
      * @throws IOException if there is an error connecting to the Graphite server
      */
     public GraphiteReporter(MetricsRegistry metricsRegistry, String host, int port, String prefix, MetricPredicate predicate) throws IOException {
-        this.tickThread = metricsRegistry.threadPools().newScheduledThreadPool(1, "graphite-reporter");
-        this.metricsRegistry = metricsRegistry;
+        super(metricsRegistry, "graphite-reporter");
         this.host = host;
         this.port = port;
         if (prefix != null) {
@@ -161,16 +165,6 @@ public class GraphiteReporter implements Runnable {
             this.prefix = "";
         }
         this.predicate = predicate;
-    }
-
-    /**
-     * Starts sending output to graphite server.
-     *
-     * @param period the period between successive displays
-     * @param unit   the time unit of {@code period}
-     */
-    public void start(long period, TimeUnit unit) {
-        tickThread.scheduleAtFixedRate(this, period, period, unit);
     }
 
     @Override
@@ -184,12 +178,16 @@ public class GraphiteReporter implements Runnable {
             printRegularMetrics(epoch);
             writer.flush();
         } catch (Exception e) {
-            log.error("Error:", e);
+            if (LOG.isDebugEnabled()) {
+                LOG.debug("Error writing to Graphite", e);
+            } else {
+                LOG.warn("Error writing to Graphite: {}", e.getMessage());
+            }
             if (writer != null) {
                 try {
                     writer.flush();
                 } catch (IOException e1) {
-                    log.error("Error while flushing writer:", e1);
+                    LOG.error("Error while flushing writer:", e1);
                 }
             }
         } finally {
@@ -197,7 +195,7 @@ public class GraphiteReporter implements Runnable {
                 try {
                     socket.close();
                 } catch (IOException e) {
-                    log.error("Error while closing socket:", e);
+                    LOG.error("Error while closing socket:", e);
                 }
             }
             writer = null;
@@ -223,7 +221,7 @@ public class GraphiteReporter implements Runnable {
                             printTimer((TimerMetric) metric, simpleName, epoch);
                         }
                     } catch (Exception ignored) {
-                        log.error("Error printing regular metrics:", ignored);
+                        LOG.error("Error printing regular metrics:", ignored);
                     }
                 }
             }
@@ -234,7 +232,7 @@ public class GraphiteReporter implements Runnable {
         try {
             writer.write(data);
         } catch (IOException e) {
-            log.error("Error sending to Graphite:", e);
+            LOG.error("Error sending to Graphite:", e);
         }
     }